fix: add missing headers

+++ client/src/actions/admin/admin-delete-comment.ts
@@ -3,13 +3,21 @@
 import { ApiResponse } from "@/interfaces/actions"
 import ApiError from "@/utils/api-error"
 import { ADMIN_DELETE_COMMENT } from "@/utils/api-urls"
+import { cookies } from "next/headers"
 
 export default async function adminDeleteComment(photoId: string, commentId: string) {
     const URL = ADMIN_DELETE_COMMENT(photoId, commentId)
+    const auth_token = cookies().get("auth_token")?.value
 
     try {
+        if (!auth_token) return ApiError(new Error("You are not authenticated."))
+
         const response = await fetch(URL, {
             method: "DELETE",
+            headers: {
+                "Content-Type": "application/json",
+                Authorization: "Bearer " + auth_token
+            },
             next: { revalidate: 60, tags: ["comment"] }
         })
